feat(AccessController): Implement endpoints to login with JWT and verify the email

+++ Models/Usuario.cs
@@ -6,6 +6,6 @@ public class Usuario
   public string Nombre { get; set; }
   public string Correo { get; set; }
   public string Password { get; set; }
-  public int Estado { get; set; }
+  public byte Estado { get; set; }
   public string Token { get; set; }
 }
